backend-fastapi/labs/B-federation-security-lab/fastapi/pyproject.toml
[build-system]
requires = ["setuptools>=75", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "b-federation-security-lab-fastapi"
version = "0.1.0"
description = "Backend-only study project for Google OIDC auth, rotating refresh tokens, and TOTP 2FA."
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14,<2.0",
  "email-validator>=2.2,<3.0",
  "fastapi>=0.115,<1.0",
  "httpx>=0.28,<1.0",
  "itsdangerous>=2.2,<3.0",
  "pydantic-settings>=2.6,<3.0",
  "PyJWT[crypto]>=2.10,<3.0",
  "pyotp>=2.9,<3.0",
  "psycopg[binary]>=3.2,<4.0",
  "redis>=5.2,<6.0",
  "SQLAlchemy>=2.0,<3.0",
  "uvicorn[standard]>=0.34,<1.0",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.14,<2.0",
  "pytest>=8.3,<9.0",
  "pytest-asyncio>=0.25,<1.0",
  "pytest-cov>=6.0,<7.0",
  "ruff>=0.8,<1.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]

[tool.mypy]
python_version = "3.12"
plugins = []
warn_unused_ignores = true
warn_redundant_casts = true
disallow_untyped_defs = true
no_implicit_optional = true
